// src/state/types/actionTypes.ts
export enum ActionTypes {
    ADD_APPOINTMENT_REQUEST = "ADD_APPOINTMENT_REQUEST",
    ADD_APPOINTMENT = "ADD_APPOINTMENT",
    GET_APPOINTMENTS_REQUEST = "GET_APPOINTMENTS_REQUEST",
    GET_APPOINTMENTS = "GET_APPOINTMENTS",
    RESET_TABLE_MODE = "RESET_TABLE_MODE",
    APPOINTMENT_READ_MODE = "FORM_READ_MODE",
    APPOINTMENT_EDITING_MODE = "FORM_EDITING_MODE",
    UPDATE_APPOINTMENT_MODE = "UPDATE_APPOINTMENT_MODE",
    UPDATE_APPOINTMENT_REQUEST = "UPDATE_APPOINTMENT_REQUEST",
    UPDATE_APPOINTMENT = "UPDATE_APPOINTMENT",
    GET_APPOINTMENT_REQUEST = "GET_APPOINTMENT_REQUEST",
    GET_APPOINTMENT = 'GET_APPOINTMENT'
}